Fix minor-frame filtering for secondary (SPE-B) FSS data
The filtering that was used to select good secondary FSS data
via SPE-B telemetry was not invariant to time or initial query
start time.  It stopped working around mid-February 2013.
A new filtering algorithm that selects data based on the times
of the FSS alpha telemetry which occur at minor_frame % 32 == 0
is used.

This commit adds a modified version of the fetch.MSIDset.interpolate
method that accepts an array of times instead of start / stop/ dt.
This should be added to fetch.